% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usertracking.R
\name{add_credentials}
\alias{add_credentials}
\title{Securely ask for Google sheet ID and username}
\usage{
add_credentials(file = ".google-sheets-credentials", overwrite = FALSE)
}
\arguments{
\item{file}{File used to store credentials. Defaults to \code{.google-sheets-credentials}.}

\item{overwrite}{Whether to overwrite file if it already exists. Default is FALSE.}
}
\value{
Nothing, used for side-effects only
}
\description{
Uses Rstudio secret functionality together with the keyring package to securely
store the input values. If keyring is not installed and up to date, the option to
do so will be given. Once keyring is present, tick the box to save the secrets.
The next time this is run, the secrets will be pre-filled.
}
\examples{
\dontrun{
add_credentials(overwrite = TRUE)
}
}
