% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usertracking.R
\name{set_credentials}
\alias{set_credentials}
\title{Set environment variables for the Google sheet ID and username}
\usage{
set_credentials(file = ".google-sheets-credentials")
}
\arguments{
\item{file}{File used to store credentials. Defaults to \code{.google-sheets-credentials}.}
}
\value{
Nothing, used for side-effects only
}
\description{
Set environment variables for the Google sheet ID and username
}
\examples{
\dontrun{
set_credentials()
}
}
