% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usertracking.R
\name{set_user_tracking}
\alias{set_user_tracking}
\title{Add user tracking}
\usage{
set_user_tracking(google_email, sheet_id, session)
}
\arguments{
\item{google_email}{Email used for Google account username.}

\item{sheet_id}{Google sheet ID.}

\item{session}{Shiny session object.}
}
\value{
Nothing; used for side effect.
}
\description{
Log session ID, username (only for Private apps), session start and end to a
Google sheet.
}
\examples{
\dontrun{
library(shiny)

ui <- fluidPage()
server <- function(input, output, session) {
  shinyusertracking::set_user_tracking(
    "joe.bloggs@google.com",
    "1234567890987654321"
  )
}

shinyApp(ui, server)
}

}
