% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usertracking.R
\name{setup_sheet}
\alias{setup_sheet}
\title{Add a new sheet for tracking to the Google sheets}
\usage{
setup_sheet(
  sheet_name = pkgload::pkg_name(),
  columns = c("login", "logout", "duration"),
  creds = ".google-sheets-credentials"
)
}
\arguments{
\item{sheet_name}{Name for the sheet. Default is to use current package name.}

\item{columns}{Which columns to log, from id, username, login, logout and
duration. By default login, logout and duration will be logged.}

\item{creds}{File used to store credentials. Defaults to \code{.google-sheets-credentials}.}
}
\value{
Nothing, used for side-effects only
}
\description{
Add a new sheet for tracking to the Google sheets
}
\examples{
\dontrun{
setup_sheet("A new Shiny app", c("login", "logout", "duration"))
}
}
