% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usertracking.R
\name{use_logging}
\alias{use_logging}
\title{Add visit tracking to Shiny app}
\usage{
use_logging(
  sheet_name = pkgload::pkg_name(),
  columns = c("login", "logout", "duration"),
  creds = ".google-sheets-credentials"
)
}
\arguments{
\item{sheet_name}{Name for the sheet. Default is to use current package name.}

\item{columns}{Which columns to log, from id, username, login, logout and
duration. By default login, logout and duration will be logged.}

\item{creds}{File used to store credentials. Defaults to \code{.google-sheets-credentials}.}
}
\value{
Nothing, used for side-effects only
}
\description{
Log session ID, username (only for Private apps), session start, end and
duration to a Google sheet.
}
\examples{
\dontrun{
library(shiny)

ui <- fluidPage()
server <- function(input, output, session) {
  shinyusertracking::set_user_tracking(
    c("login", "logout", "duration"),
    session
  )
}

shinyApp(ui, server)
}
}
