% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usertracking.R
\name{set_user_tracking}
\alias{set_user_tracking}
\title{Add user tracking}
\usage{
set_user_tracking(columns = NULL, session)
}
\arguments{
\item{columns}{Which columns to record, from id, username, login, logout and
duration. By default all will be recorded.}

\item{session}{Shiny session object.}
}
\value{
Nothing; used for side effect.
}
\description{
Log session ID, username (only for Private apps), session start, end and
duration to a Google sheet.
}
\examples{
\dontrun{
library(shiny)

ui <- fluidPage()
server <- function(input, output, session) {
  shinyusertracking::set_user_tracking(
    c("login", "logout", "duration"),
    session
  )
}

shinyApp(ui, server)
}

}
